Поместила конфиги в settings
инициализацию трейсера поместила в if + добавила в конфиг и env

auth_service/src/social_services/config.py перенесла в общий конфиг

Добавила версионирование ручек и папок в проекту

обозначила src как источник

Добавила префиксы к blueprint, но не вижу, куда добавить /v1, чтобы при открытии приложения открывался путь 127.0.0.1:5000/v1

авторизацию через яндекс и гугл проверила
