+++ .env-api-example
@@ -14,6 +14,7 @@ ELASTIC_PORT=9200
 
 API_HOST=nginx
 API_PORT=80
+ENABLE_TRACER=1
 AUTH_HOST=auth_service
 MOVIES_HOST=fastapi
