+++ README.md
@@ -33,7 +33,7 @@ poetry shell
 
 **Запустить команду в виртуальном окружении** (можно запускать извне окружения)
 ```shell
-poetry run python3 auth_service/src/main.py # пример запуска команды
+poetry run python3 auth_service/src/app.py # пример запуска команды
 ```
 
 **Установить библиотеку в виртуальном окружении** (можно запускать извне окружения)
