+++ auth_service/src/v1/social_services/google_data.py
@@ -1,4 +1,4 @@
-from src.social_services.base import BaseDataParser, SocialUserModel
+from src.v1.social_services.base import BaseDataParser, SocialUserModel
 
 
 class GoogleDataParser(BaseDataParser):
