+++ auth_service/src/v1/social_services/oauth_services.py
@@ -1,20 +1,10 @@
 from enum import Enum
 
 from authlib.integrations.flask_client import OAuth  # type: ignore
+from src.v1.core.config import settings
 
-from src.core.views import add_route
-from src.social_services.config import (
-    GOOGLE_CLIENT_ID,
-    GOOGLE_CLIENT_SECRET,
-    GOOGLE_DISCOVERY_URL,
-    GOOGLE_OAUTH_SETTINGS,
-    YANDEX_ACCESS_TOKEN_URL,
-    YANDEX_API_BASE_URL,
-    YANDEX_APP_ID,
-    YANDEX_APP_SECRET,
-    YANDEX_AUTHORIZE_URL,
-)
-from src.social_services.social_auth import (
+from src.v1.core.views import add_route
+from src.v1.social_services.social_auth import (
     social_auth_factory,
     social_login_factory,
 )
@@ -35,10 +25,10 @@ def google_register(oauth: OAuth) -> None:
     """
     oauth.register(
         name=Services.GOOGLE.value,
-        client_id=GOOGLE_CLIENT_ID,
-        client_secret=GOOGLE_CLIENT_SECRET,
-        server_metadata_url=GOOGLE_DISCOVERY_URL,
-        client_kwargs=GOOGLE_OAUTH_SETTINGS,
+        client_id=settings.GOOGLE_CLIENT_ID,
+        client_secret=settings.GOOGLE_CLIENT_SECRET,
+        server_metadata_url=settings.GOOGLE_DISCOVERY_URL,
+        client_kwargs=settings.GOOGLE_OAUTH_SETTINGS,
     )
 
 
@@ -52,11 +42,11 @@ def yandex_register(oauth: OAuth) -> None:
     """
     oauth.register(
         name=Services.YANDEX.value,
-        client_id=YANDEX_APP_ID,
-        client_secret=YANDEX_APP_SECRET,
-        api_base_url=YANDEX_API_BASE_URL,
-        access_token_url=YANDEX_ACCESS_TOKEN_URL,
-        authorize_url=YANDEX_AUTHORIZE_URL,
+        client_id=settings.YANDEX_APP_ID,
+        client_secret=settings.YANDEX_APP_SECRET,
+        api_base_url=settings.YANDEX_API_BASE_URL,
+        access_token_url=settings.YANDEX_ACCESS_TOKEN_URL,
+        authorize_url=settings.YANDEX_AUTHORIZE_URL,
     )
