+++ auth_service/src/v1/social_services/yandex_data.py
@@ -1,10 +1,10 @@
-from src.social_services.base import BaseDataParser, SocialUserModel
-from src.social_services.config import YANDEX_USERINFO_URL
+from src.v1.core.config import settings
+from src.v1.social_services.base import BaseDataParser, SocialUserModel
 
 
 class YandexDataParser(BaseDataParser):
     def get_user_info(self) -> SocialUserModel:
-        userinfo = self.client.get(YANDEX_USERINFO_URL)
+        userinfo = self.client.get(settings.YANDEX_USERINFO_URL)
         userinfo_dict = userinfo.json()
         return SocialUserModel(
             open_id=userinfo_dict.get('id'),
