+++ auth_service/src/v1/utils/template_utils.py
