+++ auth_service/templates/security/login_user.html
@@ -38,11 +38,11 @@
         </div>
         <button class="button block is-info is-large is-fullwidth">Login</button>
         <div class="block">Doesn't have an account?</div>
-        <a href="/auth/yandex">
+        <a href="/v1/auth/yandex">
           <div class="box"  style="background-color: black;color: white;">Sign in with Yandex</div>
         </a>
         <br/>
-        <a href="/auth/google">
+        <a href="/v1/auth/google">
           <div class="box" style="background-color: #4285F4;color: white;">Sign in with Google</div>
         </a>
       </form>
