+++ pyproject.toml
@@ -79,7 +79,7 @@ exclude = [
     "auth_service/src/core/config.py",
     "auth_service/src/db/postgres.py",
     "auth_service/src/core/jaeger.py",
-    "auth_service/main.py",
+    "auth_service/app.py",
     "venv",
     "postgres_data",
     "redis_data",
@@ -113,7 +113,7 @@ select = [
     "W",
 ]
 ignore = ["A003", "B006", "RET504", "N803", "PLE1205", "S105"]
-exclude = ["auth_service/main.py", "dbdata", "postgres_data", "redis_data", "venv", "tests", "etl", "fastapi-solution"]
+exclude = ["auth_service/app.py", "dbdata", "postgres_data", "redis_data", "venv", "tests", "etl", "fastapi-solution"]
 target-version = "py310"
 
 [tool.ruff.per-file-ignores]
